Moved Prometheus package to higher position.
